+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/combat/AutoProjectile.kt
@@ -22,7 +22,7 @@ object AutoProjectile : Module("AutoProjectile", Category.COMBAT) {
     // TODO: Do not throw bridge eggs
     private val facingEnemy by boolean("FacingEnemy", true)
 
-    // TODO: Turn Range into a float range value, to avoid conflicting with KillAura
+    // TODO: Turn this into a float range value, to avoid conflicting with KillAura
     private val range by float("Range", 8f, 1f..20f, suffix = "blocks")
     private val throwDelay by intRange("ThrowDelay", 1000..1500, 50..2000, suffix = "ms")